refactor: move TokenScanner to OpenApi namespace
TokenScanner is a generic PHP file scanner used by both pipelines, not
an "analyser". Moves it to OpenApi\TokenScanner and leaves a BC
subclass at the old OpenApi\Analysers\TokenScanner location.
