+++ .php-cs-fixer.dist.php
@@ -21,6 +21,8 @@ $finder = PhpCsFixer\Finder::create()
             && !strpos($file->getPathname(), 'tests/Fixtures/PHP/DocblockAndTypehintTypes.php')
             // parameter docblock for PHP 8.6
             && !strpos($file->getPathname(), 'tests/Fixtures/Scratch/Docblocks.php')
+            // FQDN in data provider
+            && !strpos($file->getPathname(), 'tests/TokenScannerTest.php')
         ;
     })
     ->in(__DIR__);
